+++ src/webpack/resolve.js
@@ -1,5 +1,6 @@
 import path from 'path'
 import Utils from '../utils/'
+
 let alias = Utils.webpack.alias
 
 export default {
@@ -9,6 +10,6 @@ export default {
         path.join(Utils.path.rootPath, 'node_modules')
     ],
     //require时候自动补全扩展名;
-    extensions: ['.js', '.jsx', '.ts', '.tsx', '.json', '.html', '.less', '.css'],
+    extensions: ['.mjs', '.js', '.jsx', '.ts', '.tsx', '.json', '.html', '.less', '.css'],
     alias: alias // 别名
 }